+++ src/angular/builder/writer.rs
@@ -1,74 +1,56 @@
-use std::{
-	fs,
-	path::{Path, PathBuf},
-};
+use std::{fs, path::Path};
 
 use serde_json::json;
 
-use crate::{ChapterWithCodeBlocks, Config, Context, Result};
+use crate::{codeblock::CodeBlock, Config, Context, Result};
 
-pub(super) struct PostBuildReplacement {
-	pub(super) chapter_path: PathBuf,
-	pub(super) has_playgrounds: bool,
-	pub(super) project_folder: String,
-	pub(super) script_basename: String,
-	pub(super) script_marker: String,
-	#[cfg_attr(not(all(unix, feature = "background")), allow(dead_code))]
-	pub(super) made_changes_to_scripts: bool,
-}
-
-pub(super) enum Writer {
-	Default,
-	#[cfg_attr(not(all(unix, feature = "background")), allow(dead_code))]
-	ChangedOnly,
+pub(super) struct Writer {
+	changed_only: bool,
+	chapter_to_angular_file: Vec<(String, String)>,
 }
 
 impl Writer {
-	pub(super) fn write<P: AsRef<Path>>(&self, path: P, contents: &str) -> Result<bool> {
-		if matches!(self, Self::ChangedOnly)
+	pub(super) fn new(changed_only: bool) -> Self {
+		Self {
+			changed_only,
+			chapter_to_angular_file: Vec::new(),
+		}
+	}
+
+	pub(super) fn write<P: AsRef<Path>>(&self, path: P, contents: &str) -> Result<()> {
+		if self.changed_only
 			&& matches!(fs::read_to_string(&path), Ok(existing) if existing.eq(contents))
 		{
-			return Ok(false);
+			return Ok(());
 		}
 
 		fs::write(&path, contents)?;
 
-		Ok(true)
+		Ok(())
+	}
+
+	#[allow(clippy::unused_self)]
+	pub(super) fn write_force<P: AsRef<Path>>(&self, path: P, contents: &str) -> Result<()> {
+		fs::write(&path, contents)?;
+
+		Ok(())
 	}
 
 	pub(super) fn write_chapter<P: AsRef<Path>>(
-		&self,
-		config: &Config,
+		&mut self,
 		root: P,
 		index: usize,
-		chapter: ChapterWithCodeBlocks,
-	) -> Result<PostBuildReplacement> {
+		chapter_path: &Path,
+		code_blocks: Vec<CodeBlock>,
+	) -> Result<()> {
 		let root = root.as_ref();
 		let project_folder = format!("code_{index}");
-		let script_marker = chapter.script_marker().to_owned();
-		let has_playgrounds = chapter.has_playgrounds();
 
 		let absolute_project_folder = root.join(&project_folder);
-		let mut made_changes_to_scripts = false;
-
-		if matches!(self, Self::Default) && absolute_project_folder.exists() {
-			fs::remove_dir_all(&absolute_project_folder)
-				.context("failed to clear project folder")?;
-			made_changes_to_scripts = true;
-		}
 
 		fs::create_dir_all(&absolute_project_folder).context("failed to create project folder")?;
 
-		let mut main_script =
-			Vec::with_capacity(2 + config.polyfills.len() + chapter.number_of_code_blocks());
-
-		if !config.polyfills.contains(&"zone.js".to_owned()) {
-			main_script.push("import 'zone.js';".to_owned());
-		}
-
-		for polyfill in &config.polyfills {
-			main_script.push(format!("import '{polyfill}';"));
-		}
+		let mut main_script = Vec::with_capacity(1 + code_blocks.len());
 
 		main_script.push(
 			"\n\
@@ -84,19 +66,12 @@ impl Writer {
 			.to_owned(),
 		);
 
-		let chapter_path = chapter.source_path().to_owned();
-
-		for (code_block_index, code_block) in chapter.into_iter().enumerate() {
-			let changed_script = self
-				.write(
-					absolute_project_folder.join(&format!("codeblock_{code_block_index}.ts")),
-					&code_block.code_to_run,
-				)
-				.context("failed to write code block")?;
-
-			if changed_script && !made_changes_to_scripts {
-				made_changes_to_scripts = changed_script;
-			}
+		for (code_block_index, code_block) in code_blocks.into_iter().enumerate() {
+			self.write(
+				absolute_project_folder.join(format!("codeblock_{code_block_index}.ts")),
+				&code_block.code_to_run,
+			)
+			.context("failed to write code block")?;
 
 			main_script.push(format!(
 				"\
@@ -109,18 +84,49 @@ impl Writer {
 
 		let script_basename = project_folder.clone();
 
-		let angular_main = format!("{}/{}.ts", &project_folder, &script_basename);
-		self.write(root.join(angular_main), &main_script.join("\n"))
-			.context("failed to write main chapter import")?;
+		let angular_main = format!("./{}/{}", &project_folder, &script_basename);
+		self.write(
+			root.join(format!("{angular_main}.ts")),
+			&main_script.join("\n"),
+		)
+		.context("failed to write main chapter import")?;
 
-		Ok(PostBuildReplacement {
-			chapter_path,
-			has_playgrounds,
-			project_folder,
-			script_basename,
-			script_marker,
-			made_changes_to_scripts,
-		})
+		self.chapter_to_angular_file.push((
+			chapter_path.to_string_lossy().into_owned(),
+			format!("{angular_main}.js"),
+		));
+
+		Ok(())
+	}
+
+	pub(super) fn write_main<P: AsRef<Path>>(&self, config: &Config, root: P) -> Result<()> {
+		let mut main_script =
+			Vec::with_capacity(3 + config.polyfills.len() + self.chapter_to_angular_file.len());
+
+		if !config.polyfills.contains(&"zone.js".to_owned()) {
+			main_script.push("import 'zone.js';".to_owned());
+		}
+
+		for polyfill in &config.polyfills {
+			main_script.push(format!("import '{polyfill}';"));
+		}
+
+		main_script.push("\nconst mods = new Map([".to_owned());
+
+		for (chapter_path, relative_script_path) in &self.chapter_to_angular_file {
+			main_script.push(format!(
+				r#"  ["{chapter_path}", () => import("{relative_script_path}")],"#,
+			));
+		}
+
+		main_script.push("]);\n\nmods.get((document.querySelector('#load-angular') as HTMLElement)?.dataset.path as string)?.();".to_owned());
+
+		self.write_force(
+			root.as_ref().join("load-angular.ts"),
+			&main_script.join("\n"),
+		)?;
+
+		Ok(())
 	}
 
 	pub(super) fn write_tsconfig(&self, config: &Config) -> Result<()> {
@@ -147,45 +153,6 @@ impl Writer {
 			&serde_json::to_string(&tsconfig)?,
 		)
 		.context("failed to write tsconfig.json")?;
-
-		Ok(())
-	}
-}
-
-pub(super) struct PlaygroundScriptWriter<'a> {
-	config: &'a Config,
-	has_playgrounds: bool,
-}
-
-impl<'a> PlaygroundScriptWriter<'a> {
-	pub(super) fn new(config: &Config) -> PlaygroundScriptWriter<'_> {
-		PlaygroundScriptWriter {
-			config,
-			has_playgrounds: false,
-		}
-	}
-
-	pub(super) fn insert_playground_script(
-		&mut self,
-		replacement: &PostBuildReplacement,
-		chapter: &mut String,
-		path_to_root: &str,
-	) {
-		if replacement.has_playgrounds {
-			chapter.push_str(&format!(
-				r#"<script type="module" src="{path_to_root}/playground-io.min.js"></script>"#
-			));
-			self.has_playgrounds = true;
-		}
-	}
-
-	pub(super) fn write_playground_file(&self) -> Result<()> {
-		if self.has_playgrounds {
-			fs::write(
-				self.config.target_folder.join("playground-io.min.js"),
-				crate::js::PLAYGROUND_SCRIPT,
-			)?;
-		}
 
 		Ok(())
 	}