+++ src/js/mod.rs
@@ -1,8 +1,1 @@
 pub(crate) static PLAYGROUND_SCRIPT: &str = include_str!("playground-io.min.js");
-
-pub(crate) static EXPERIMENTAL_BUILDER_IMPLEMENTATION: &str =
-	include_str!("experimental-builder/builder.min.mjs");
-pub(crate) static EXPERIMENTAL_BUILDER_MANIFEST: &str =
-	include_str!("experimental-builder/package.json");
-pub(crate) static EXPERIMENTAL_BUILDER_SCHEMA: &str =
-	include_str!("experimental-builder/schema.json");